Fix IntelliJ being a dum dum
For some reason on this new update the guide line for
80 characters max on a git commit disappeared.
Now you have to do this weird thing to fix it.

Also I've learned that IntelliJ has the WORST
preferences functionality. Everything has to be project-level
so it's either a new project default, or you have to set preferences
on every single project you have.

Why can't it just be global preferences and project preferences,
where project preferences override global preferences? Anyway,
because of this preferences mess, you have to have this thingy
in your project when I just want my guide lines back. Blegh.